requirements.txt
openai
python-dotenv
flask
flask-cors
PyPDF2
python-docx
pytesseract
Pillow
